feat: optimize code

+++ src/Worker/AbstractMainProcess.php
@@ -45,13 +45,13 @@ abstract class AbstractMainProcess extends AbstractProcess
 
         if (function_exists('customLoadWorkerConf')) {
             $workerConfList = customLoadWorkerConf();
-        }else if(defined('WORKER_CONF_FILE')) {
+        } else if (defined('WORKER_CONF_FILE')) {
             $mainManager = \Swoolefy\Worker\MainManager::getInstance();
             $workerConfList = \Swoolefy\Worker\MainManager::loadWorkerConf(WORKER_CONF_FILE);
         }
 
         $workerConfListNew = [];
-        if(!empty($workerConfList)) {
+        if (!empty($workerConfList)) {
             // Specify Process to Run When dev or test to debug, Avoid the impact of other processes
             $onlyProcess = Helper::getCliParams('only');
             if ($onlyProcess) {
@@ -68,13 +68,15 @@ abstract class AbstractMainProcess extends AbstractProcess
 
                 if (empty($workerConfListNew)) {
                     fmtPrintError("Not Found Specify Process --only={$onlyProcess}, All Process Exited!");
-                    $masterPid = $mainManager->getMasterPid();
-                    // kill master to exit
-                    \Swoole\Process::kill($masterPid, SIGTERM);
-                }else {
+                    if (isset($mainManager)) {
+                        $masterPid = $mainManager->getMasterPid();
+                        // kill master to exit
+                        \Swoole\Process::kill($masterPid, SIGTERM);
+                    }
+                } else {
                     $workerConf = $workerConfListNew;
                 }
-            }else {
+            } else {
                 $workerConf = $workerConfList;
             }
         }
